Workspace roles, review gating, and chat alignment
Role-correct uploads and review actions:
- FilesTab splits upload UX by role: editors upload deliverable review cuts (video-only), clients upload reference material (broader file types). Delete is symmetry-gated client-side and on the server.
- VideoReviewModal: Resolve/Re-open is editor-only. Clients see a static Open/Resolved status pill instead. Switched from Redux user.role to summary.role passed through FilesTab.
- projectFile.controller.ts: assertScopeAccess returns the caller role; createProjectFile and deleteProjectFile reject category mismatches per role.
- videoReview.controller.ts: setResolved returns 403 when caller is the client.

Block completion until comments are resolved:
- New countOpenReviewComments helper in videoReview.controller.ts.
- workspace.controller.ts adds openReviewCount to job and order summaries.
- timeline.routes.ts blocks COMPLETED milestone transitions on jobs while open comments > 0 (returns 409).
- delivery.controller.ts gates submitFinalDelivery and deliverFinalMaster on the same check for both JOB and ORDER scopes.
- MilestonesTab / RevisionsTab / DeliveryPanel surface an amber "N open comment" pill, disable the relevant action, and intercept clicks with a clear toast.

Chat alignment and role styling:
- Workspace/utils.js gains bubbleClass({mine,role}) and roleCaption({mine,role}) helpers (indigo for me, emerald for editor, sky for client).
- ChatRail OrderMessagePanel: robust isMine across senderId/sender.id/sender_id/from/author.id; flex-row-reverse plus avatar-on-mine; role caption above bubble.
- ChatPanel: same alignment and theming; resolveSenderId widened to also try from and author.id.
- message.controller.ts: HTTP response and broadcast payloads always emit numeric senderId.
- chatStore.js: optimistic messages set numeric senderId; new normalizeServerMessage helper preserves it through reconciliation.

+++ vid/prisma/migrations/20260422_revenue_streams.sql
@@ -1,17 +1,23 @@
 -- =============================================
 -- REVENUE: Service Fee / Platform Commission
 -- =============================================
+-- NOTE: amount columns are NUMERIC(12,2) to match Order."totalPrice" and the
+-- cent-precise output of pricing.service.ts (roundMoney). Earlier versions of
+-- this migration used INTEGER which silently rejected fees like 11.13 with
+-- "invalid input syntax for type integer". The reconcile script
+-- scripts/fix-order-fee-columns.mjs upgrades pre-existing INTEGER columns
+-- in place.
 ALTER TABLE "Order" ADD COLUMN IF NOT EXISTS "platformFeePercent" DOUBLE PRECISION DEFAULT 12.5;
-ALTER TABLE "Order" ADD COLUMN IF NOT EXISTS "platformFeeAmount" INTEGER DEFAULT 0;
-ALTER TABLE "Order" ADD COLUMN IF NOT EXISTS "clientFeePercent" DOUBLE PRECISION DEFAULT 3.5;
-ALTER TABLE "Order" ADD COLUMN IF NOT EXISTS "clientFeeAmount" INTEGER DEFAULT 0;
-ALTER TABLE "Order" ADD COLUMN IF NOT EXISTS "freelancerPayout" INTEGER DEFAULT 0;
+ALTER TABLE "Order" ADD COLUMN IF NOT EXISTS "platformFeeAmount"  NUMERIC(12,2)    DEFAULT 0;
+ALTER TABLE "Order" ADD COLUMN IF NOT EXISTS "clientFeePercent"   DOUBLE PRECISION DEFAULT 3.5;
+ALTER TABLE "Order" ADD COLUMN IF NOT EXISTS "clientFeeAmount"    NUMERIC(12,2)    DEFAULT 0;
+ALTER TABLE "Order" ADD COLUMN IF NOT EXISTS "freelancerPayout"   NUMERIC(12,2)    DEFAULT 0;
 
 -- =============================================
 -- REVENUE: Template Marketplace Commission
 -- =============================================
-ALTER TABLE "TemplatePurchase" ADD COLUMN IF NOT EXISTS "platformCommission" INTEGER DEFAULT 0;
-ALTER TABLE "TemplatePurchase" ADD COLUMN IF NOT EXISTS "sellerPayout" INTEGER DEFAULT 0;
+ALTER TABLE "TemplatePurchase" ADD COLUMN IF NOT EXISTS "platformCommission" NUMERIC(12,2) DEFAULT 0;
+ALTER TABLE "TemplatePurchase" ADD COLUMN IF NOT EXISTS "sellerPayout"       NUMERIC(12,2) DEFAULT 0;
 
 -- =============================================
 -- REVENUE: Premium Subscriptions
@@ -96,7 +102,7 @@ CREATE INDEX IF NOT EXISTS "idx_enterprise_member" ON "EnterpriseMember" ("userI
 CREATE TABLE IF NOT EXISTS "PlatformRevenue" (
   "id"                SERIAL PRIMARY KEY,
   "type"              VARCHAR(30) NOT NULL,
-  "amount"            INTEGER NOT NULL DEFAULT 0,
+  "amount"            NUMERIC(12,2) NOT NULL DEFAULT 0,
   "sourceId"          INTEGER,
   "sourceType"        VARCHAR(30),
   "description"       TEXT,
